backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-filter>=23.5
drf-spectacular>=0.27.0
celery>=5.3.0
redis>=5.0.0
mysqlclient>=2.2.0
Pillow>=10.1.0
python-dotenv>=1.0.0
requests>=2.31.0
openpyxl>=3.1.0
pandas>=2.1.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pycryptodome>=3.20.0
gunicorn>=21.2.0
sentry-sdk>=1.39.0
